Move suspend next to patches
It makes more sense to me to define this next to where we setup patching
